+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## v0.2.5 - ?
 ### Added:
  - Negotiations: invited, view, comments for employer, send a message
+ - Suggests: educational_institutions
 
 ## v0.2.4 -  2017-05-01
 ### Added: